summary-model/pyproject.toml
[tool.poetry]
name = "summary-model"
version = "0.1.0"
description = ""
authors = ["david-slatinek"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
transformers = "^4.28.1"
torch = "2.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
